pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.0",
    "inquirerpy>=0.3.4",
    "python-binance>=1.0.29",
    "python-dotenv>=1.1.1",
    "rich>=14.0.0",
]
